pet_clinic_insurance_service/requirements.txt
boto3
django
psycopg2
django_rest_framework
py_eureka_client
django-health-check